requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
psycopg2-binary==2.9.9
twilio==9.2.3
transformers==4.40.0
torch==2.3.0
huggingface_hub==0.23.0
python-multipart==0.0.9
dateparser==1.3.0
APScheduler==3.11.2
pydantic-settings==2.13.1
